# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: MaintainIQ API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /issues:
    get:
      operationId: listIssues
      tags: [issues]
      summary: List issues (admin sees all, technician sees assigned)
      security: [{ bearerAuth: [] }]
      parameters:
        - name: search
          in: query
          schema: { type: string }
        - name: status
          in: query
          schema: { type: string }
        - name: priority
          in: query
          schema: { type: string }
        - name: category
          in: query
          schema: { type: string }
        - name: assetId
          in: query
          schema: { type: integer }
        - name: technicianId
          in: query
          schema: { type: integer }
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: pageSize
          in: query
          schema: { type: integer, default: 20 }
      responses:
        "200":
          description: Issue list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IssueListResponse"
  /issues/{id}:
    get:
      operationId: getIssue
      tags: [issues]
      summary: Get an issue's detail
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Issue detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Issue"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /issues/{id}/assign:
    patch:
      operationId: assignIssue
      tags: [issues]
      summary: Assign an issue to a technician (admin only)
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IssueAssignInput"
      responses:
        "200":
          description: Updated issue
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Issue"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /issues/{id}/status:
    patch:
      operationId: updateIssueStatus
      tags: [issues]
      summary: Transition an issue's status (admin or assigned technician)
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IssueStatusUpdate"
      responses:
        "200":
          description: Updated issue
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Issue"
        "400":
          description: Invalid transition
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /issues/{id}/history:
    get:
      operationId: getIssueHistory
      tags: [issues]
      summary: Get an issue's history trail
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: History entries
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HistoryListResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"

  # ---------- Maintenance ----------
  /maintenance:
    get:
      operationId: listMaintenance
      tags: [maintenance]
      summary: List maintenance work log entries for an issue
      security: [{ bearerAuth: [] }]
      parameters:
        - name: issueId
          in: query
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Maintenance entries
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MaintenanceListResponse"
  /issues/{id}/maintenance:
    post:
      operationId: createMaintenance
      tags: [maintenance]
      summary: Add a maintenance work log entry (assigned technician or admin)
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MaintenanceInput"
      responses:
        "201":
          description: Maintenance entry created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Maintenance"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"

  # ---------- History ----------
...
  /notifications:
    get:
      operationId: listNotifications
      tags: [notifications]
      summary: List notifications for the current user
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Notification list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationListResponse"
  /notifications/{id}/read:
    patch:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark a notification as read
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Updated notification
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"

  # ---------- Storage ----------
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema:
        type: integer
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string

    # ---------- Users / Auth ----------
...
    Issue:
      type: object
      required:
        [
          id,
          issueNumber,
          assetId,
          assetName,
          assetCode,
          reporterName,
          reporterEmail,
          title,
          description,
          priority,
          category,
          imageUrls,
          status,
          aiPossibleCauses,
          aiDiagnosticChecks,
          createdAt,
          updatedAt,
        ]
      properties:
        id: { type: integer }
        issueNumber: { type: string }
        assetId: { type: integer }
        assetName: { type: string }
        assetCode: { type: string }
        reporterName: { type: string }
        reporterEmail: { type: string }
        reporterPhone: { type: ["string", "null"] }
        title: { type: string }
        description: { type: string }
        priority: { type: string, enum: [low, medium, high, critical] }
        category: { type: string }
        imageUrls:
          type: array
          items: { type: string }
        status:
          type: string
          enum:
            [
              reported,
              assigned,
              inspection_started,
              maintenance_in_progress,
              waiting_for_parts,
              resolved,
              closed,
              reopened,
            ]
        assignedTechnicianId: { type: ["integer", "null"] }
        assignedTechnicianName: { type: ["string", "null"] }
        aiTitle: { type: ["string", "null"] }
        aiCategory: { type: ["string", "null"] }
        aiPriority: { type: ["string", "null"] }
        aiPossibleCauses:
          type: array
          items: { type: string }
        aiDiagnosticChecks:
          type: array
          items: { type: string }
        aiSafetyNotes: { type: ["string", "null"] }
        aiRecurringWarning: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...
    IssueListResponse:
      type: object
      required: [items, total, page, pageSize]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Issue"
        total: { type: integer }
        page: { type: integer }
        pageSize: { type: integer }
    IssueAssignInput:
      type: object
      required: [technicianId]
      properties:
        technicianId: { type: integer }
    IssueStatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum:
            [
              reported,
              assigned,
              inspection_started,
              maintenance_in_progress,
              waiting_for_parts,
              resolved,
              closed,
              reopened,
            ]
        notes: { type: string }

    # ---------- Maintenance ----------
    Maintenance:
      type: object
      required:
        [
          id,
          issueId,
          technicianId,
          technicianName,
          notes,
          replacementParts,
          images,
          createdAt,
        ]
      properties:
        id: { type: integer }
        issueId: { type: integer }
        technicianId: { type: integer }
        technicianName: { type: string }
        notes: { type: string }
        cost: { type: ["number", "null"] }
        replacementParts:
          type: array
          items: { type: string }
        timeSpentMinutes: { type: ["integer", "null"] }
        images:
          type: array
          items: { type: string }
        createdAt: { type: string, format: date-time }
    MaintenanceInput:
      type: object
      required: [notes]
      properties:
        notes: { type: string, minLength: 1 }
        cost: { type: number }
        replacementParts:
          type: array
          items: { type: string }
        timeSpentMinutes: { type: integer }
        images:
          type: array
          items: { type: string }
    MaintenanceListResponse:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Maintenance"

    # ---------- History ----------
    HistoryEntry:
      type: object
      required: [id, assetId, assetName, userName, action, createdAt]
      properties:
        id: { type: integer }
        assetId: { type: integer }
        assetName: { type: string }
        issueId: { type: ["integer", "null"] }
        issueNumber: { type: ["string", "null"] }
        userId: { type: ["integer", "null"] }
        userName: { type: string }
        action: { type: string }
        status: { type: ["string", "null"] }
        notes: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
    HistoryListResponse:
      type: object
      required: [items, total, page, pageSize]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/HistoryEntry"
        total: { type: integer }
        page: { type: integer }
        pageSize: { type: integer }

    # ---------- Dashboard ----------
...
    Notification:
      type: object
      required: [id, type, title, message, read, createdAt]
      properties:
        id: { type: integer }
        type: { type: string }
        title: { type: string }
        message: { type: string }
        read: { type: boolean }
        createdAt: { type: string, format: date-time }
    NotificationListResponse:
      type: object
      required: [items, total]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Notification"
        total: { type: integer }

    # ---------- Storage ----------
...