# lib/api-zod/src/generated/types/maintenanceInput.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * MaintainIQ API specification
 * OpenAPI spec version: 0.1.0
 */

export interface MaintenanceInput {
  /** @minLength 1 */
  notes: string;
  cost?: number;
  replacementParts?: string[];
  timeSpentMinutes?: number;
  images?: string[];
}